requirements.txt
selenium==4.10.0
undetected-chromedriver
webdriver-manager
datetime
regex
discord==2.3.2
discord.py==2.3.2
python-dotenv==1.0.1
seleniumbase